@@ -81,7 +81,8 @@ export const getXuiNodeMiddleware = () => {
       '/api/role-access/roles/getJudicialUsers',
       '/workallocation/getJudicialUsers',
       '/workallocation/caseworker/getUsersByServiceName',
-      '/api/prd/judicial/searchJudicialUserByPersonalCodes'
+      '/api/prd/judicial/searchJudicialUserByPersonalCodes',
+      '/api/prd/judicial/searchJudicialUserByIdamId'
     ],
     scope: 'openid profile roles manage-user create-user search-user',
     userName
