@@ -2,14 +2,15 @@ export const SERVICES_COURT_TYPE_MAPPINGS = {
   PROBATE: ['18'], // 18: Probate
   EMPLOYMENT: ['17'], // 17: Employment Tribunal
   IA: ['4', '23'], // 4: Asylum Support Appeals 23: Immigration and Asylum Tribunal
-  SSCS: ['31'], // 31: Social Security and Child Support Tribunal
   CIVIL: ['10'],
   PRIVATELAW: ['18'],
   PUBLICLAW: ['18'],
   BHA1: ['17'], // Employment Claims
   BFA1: ['4', '23'], // 4: Asylum Support Appeals, 23: Immigration and Asylum Tribunal
   BBA1: ['4'], // Asylum Support
+  ST_CIC: ['14'], // Special Tribunals
   BBA2: ['14'], // Criminal Injuries Compensation
+  SSCS: ['31'], // 31: Social Security and Child Support Tribunal
   BBA3: ['31'], // Social Security and Child Support Tribunal
   ABA1: ['18'], // Divorce,
   ABA2: ['18'], // Financial Remedy
