@@ -2,6 +2,7 @@ import { test, expect } from '@playwright/test';
 import { signIn } from './steps/login-steps';
 import { clickOnMainMenu } from './steps/steps-functions';
 import { retryAction } from './steps/retry-steps';
+import axeTest from "./helpers/accessibilityTestHelper";
 
 test('Search from menu 16-digit find control', async ({ page }) => {
   await signIn(page, 'IAC_CaseOfficer_R2');
@@ -12,9 +13,10 @@ test('Search from menu 16-digit find control', async ({ page }) => {
   await page.getByLabel('-digit case reference:').click();
   await page.getByLabel('-digit case reference:').fill(caseId); 
   await retryAction(async () => {
-    await page.getByRole('button', { name: 'Find' }).click();
+    await page.locator('//button[contains(text(), "Find")]').click();
     await expect(page.getByRole('heading', { name: 'Current progress of the case' })).toBeVisible();
   });
+  await axeTest(page);
   
   console.log('Check the case details are displayed');
   await expect(page.getByRole('heading', { name: 'Current progress of the case' })).toBeVisible();
@@ -39,6 +41,7 @@ test('Search from page Search', async ({ page }) => {
   await page.getByLabel('16-digit case reference', { exact: true }).fill(nonExistentCaseId);
   await page.getByRole('button', { name: 'Search' }).click();
   await expect(page.getByRole('heading', { name: 'No results found' })).toBeVisible();
+  await axeTest(page);
 });
 
 function findCaseId(page: any) {
