@@ -9,6 +9,8 @@ import { CaseReferencePipe } from '../../../hearings/pipes/case-reference.pipe';
 import { AllocateRoleService } from '../../../role-access/services';
 import { CaseworkerDataService, WASupportedJurisdictionsService } from '../../../work-allocation/services';
 import { RestrictedCaseAccessContainerComponent } from './restricted-case-access-container.component';
+import { HttpClient, HttpHandler } from '@angular/common/http';
+import { JudicialRefDataService } from 'src/hearings/services/judicial-ref-data.service';
 
 @Pipe({ name: 'rpxTranslate' })
 class RpxTranslateMockPipe implements PipeTransform {
@@ -24,6 +26,7 @@ describe('RestrictedCaseAccessContainerComponent', () => {
   const mockWASupportedJurisdictionsService = jasmine.createSpyObj('WASupportedJurisdictionsService', ['getWASupportedJurisdictions']);
   const mockCaseworkerDataService = jasmine.createSpyObj('CaseworkerDataService', ['getUsersFromServices']);
   const mockLoadingService = jasmine.createSpyObj('LoadingService', ['register', 'unregister']);
+  const mockJudicialRefDataService = jasmine.createSpyObj('JudicialRefDataService', ['searchJudicialUserByIdamID']);
   const mockActivatedRoute = {
     snapshot: {
       params: {
@@ -31,6 +34,32 @@ describe('RestrictedCaseAccessContainerComponent', () => {
       }
     }
   };
+  const firstJudgeResponse =[
+    {
+      title: 'Ms',
+      knownAs: 'jane',
+      surname: 'doe',
+      fullName: 'Ms Jane Doe',
+      emailId: 'jane.doe@imajudge.net',
+      idamId: '6343',
+      initials: 'JD',
+      postNominals: null,
+      personalCode: '1234568'
+    }
+  ];
+  const secondJudgeResponse =[
+    {
+      title: 'Mr',
+      knownAs: 'bob',
+      surname: 'Smith',
+      fullName: 'Mr Bob Smith',
+      emailId: 'bob.smith@imajudge.net',
+      idamId: '12343',
+      initials: 'BS',
+      postNominals: null,
+      personalCode: '1234568'
+    }
+  ];
   const mockRouter = {
     navigate: jasmine.createSpy()
   };
@@ -49,7 +78,10 @@ describe('RestrictedCaseAccessContainerComponent', () => {
         { provide: CaseworkerDataService, useValue: mockCaseworkerDataService },
         { provide: LoadingService, useValue: mockLoadingService },
         { provide: ActivatedRoute, useValue: mockActivatedRoute },
-        { provide: Router, useValue: mockRouter }
+        { provide: Router, useValue: mockRouter },
+        { provide: JudicialRefDataService, useValue: mockJudicialRefDataService },
+        HttpClient,
+        HttpHandler
       ]
     })
       .compileComponents();
@@ -58,6 +90,15 @@ describe('RestrictedCaseAccessContainerComponent', () => {
     mockCaseworkerDataService.getUsersFromServices.and.returnValue(of([CASEWORKERS.JANE_DOE, CASEWORKERS.JOHN_SMITH]));
     mockLoadingService.register.and.callThrough();
     mockLoadingService.unregister.and.callThrough();
+    let callCount = 0;
+    mockJudicialRefDataService.searchJudicialUserByIdamID.and.callFake(() => {
+      callCount++;
+      if (callCount === 1) {
+        return of(firstJudgeResponse);
+      } else if (callCount === 2) {
+        return of(secondJudgeResponse);
+      }
+    });
     fixture = TestBed.createComponent(RestrictedCaseAccessContainerComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
@@ -70,6 +111,12 @@ describe('RestrictedCaseAccessContainerComponent', () => {
     expect(mockWASupportedJurisdictionsService.getWASupportedJurisdictions).toHaveBeenCalled();
     expect(mockCaseworkerDataService.getUsersFromServices).toHaveBeenCalled();
     expect(mockLoadingService.unregister).toHaveBeenCalled();
+    expect(mockJudicialRefDataService.searchJudicialUserByIdamID).toHaveBeenCalled();
+    expect(component.restrictedCases).toEqual([
+      { user: 'Ms Jane Doe', email: 'jane.doe@imajudge.net', role: 'Lead judge' },
+      { user: 'Mr Bob Smith', email: 'bob.smith@imajudge.net', role: 'Lead judge' },
+      { user: 'John Smith', email: 'john.smith@caseworkers.gov.uk', role: 'Case manager' }
+    ]);
   });
 
   it('should unsubscribe', () => {
