@@ -142,6 +142,8 @@ export class HearingActualsAddEditSummaryComponent extends HearingActualsSummary
 
   public haveParticipantsBeenAdded(hearingDay: ActualHearingDayModel): boolean {
     const individualParties = this.individualParties;
-    return individualParties.length !== hearingDay.actualDayParties.length;
+    if (hearingDay && this.individualParties){
+      return individualParties?.length !== hearingDay?.actualDayParties?.length;
+    }
   }
 }
