@@ -251,7 +251,7 @@ describe('HearingTimingSectionComponent', () => {
     });
     component.onChange('hearingSpecificDate');
     expect(component.changeEditHearing.emit).toHaveBeenCalledWith({
-      fragmentId: 'hearingSpecificDate', changeLink: '/hearings/request/hearing-timing#noSpecificDate'
+      fragmentId: 'hearingSpecificDate', changeLink: '/hearings/request/hearing-timing#hearingDateRange'
     });
     component.onChange('hearingPriority');
     expect(component.changeEditHearing.emit).toHaveBeenCalledWith({
@@ -358,23 +358,6 @@ describe('HearingTimingSectionComponent', () => {
       component.ngOnInit();
       expect(component.hearingDateChanged).toEqual(false);
     });
-
-    it('should return false if no hearing window', () => {
-      component.hearingRequestToCompareMainModel = {
-        ...initialState.hearings.hearingRequestToCompare.hearingRequestMainModel,
-        hearingDetails: {
-          ...initialState.hearings.hearingRequestToCompare.hearingRequestMainModel.hearingDetails,
-          hearingWindow: null
-        }
-      };
-      component.serviceHearingValuesModel = {
-        ...serviceHearingValuesModel,
-        hearingWindow: {}
-      };
-
-      component.ngOnInit();
-      expect(component.hearingDateChanged).toEqual(false);
-    });
   });
 
   it('should return true if firstDateTimeMustBe is removed', () => {
@@ -399,5 +382,36 @@ describe('HearingTimingSectionComponent', () => {
     };
     component.ngOnInit();
     expect(component.hearingDateChanged).toEqual(true);
+  });
+
+  describe('testing the creation of the change url', () => {
+    it('should return the correct URL fragment when earliestHearingDate or latestHearingDate is set', () => {
+      component.earliestHearingDate = '2023-10-01';
+      component.latestHearingDate = '2023-10-10';
+      // eslint-disable-next-line dot-notation
+      const result = component['setDateSectionForChange']();
+
+      expect(result).toBe('/hearings/request/hearing-timing#hearingDateRange');
+    });
+
+    it('should return the correct URL fragment when firstHearingDate is set', () => {
+      component.earliestHearingDate = '';
+      component.latestHearingDate = '';
+      component.firstHearingDate = '2023-10-01';
+      // eslint-disable-next-line dot-notation
+      const result = component['setDateSectionForChange']();
+
+      expect(result).toBe('/hearings/request/hearing-timing#hearingSingleDate');
+    });
+
+    it('should return the default URL fragment when no dates are set', () => {
+      component.earliestHearingDate = '';
+      component.latestHearingDate = '';
+      component.firstHearingDate = '';
+      // eslint-disable-next-line dot-notation
+      const result = component['setDateSectionForChange']();
+
+      expect(result).toBe('/hearings/request/hearing-timing#noSpecificDate');
+    });
   });
 });
